requirements.txt
aio-pika==9.4.3
aiormq==6.8.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
confluent-kafka==2.6.0
coverage==7.6.5
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
Faker==33.0.0
fastapi==0.115.4
fastapi-cli==0.0.5
filelock==3.16.1
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.8.2
lazy-model==0.2.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
motor==3.6.0
multidict==6.1.0
openai==1.57.4
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pluggy==1.5.0
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.2
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.2
toml==0.10.2
tqdm==4.67.1
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
